requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
boto3==1.37.23
botocore==1.37.23
certifi==2025.1.31
colorama==0.4.6
deprecation==2.1.0
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
multidict==6.3.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic-bin==0.4.14
realtime==2.4.2
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
